bug fixed

+++ app/Http/Controllers/Administrator/EmployeeController.php
@@ -10,6 +10,7 @@ use App\Services\Domain\OrgService;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\MessageBag;
+use Image;
 
 class EmployeeController extends Controller
 {
@@ -34,13 +35,24 @@ class EmployeeController extends Controller
     public function create(Request $request, EmployeeService $employeeService, OrgService $orgService, Organization $org)
     {
         if ($request->method() == 'POST') {
-            $request->merge(['org' => $org]);
-            $request->validate([
+            $validation = [
+                'code' => 'required',
                 'name' => 'required',
+                'identity_number' => 'required',
                 'school' => 'required',
-                'org' => 'required',
                 'gender' => 'in:' . Employee::GENDER_MALE . ',' . Employee::GENDER_FEMALE,
                 'dateOfBirth' => 'required|date_format:"d-m-Y',
+                'photo' => 'mimes:jpeg,jpg,png,bmp|max:540'
+            ];
+
+            $request->validate($validation, [], [
+                'code' => ucfirst(trans('common.code')),
+                'name' => ucfirst(trans('common.name')),
+                'identity_number' => ucfirst(trans('common.identity_number')),
+                'school' => ucfirst(trans('common.school')),
+                'gender' => ucfirst(trans('common.gender')),
+                'dateOfBirth' => ucfirst(trans('common.date_of_birth')),
+                'photo' => ucfirst(trans('common.photo')),
             ]);
 
             $messageBag = new MessageBag;
@@ -50,12 +62,23 @@ class EmployeeController extends Controller
                 $school = $orgService->findById($request->get('school'));
             }
             if (!$school) {
-                $messageBag->add('school', trans('common.invalid_institute'));
+                $messageBag->add('school', trans('common.invalid_school'));
                 return redirect()->route('administrator.employee.create', ['org' => $org->getId()])->withErrors($messageBag);
             }
 
             try {
                 $requestData = $request->all();
+
+                if ($request->hasFile('photo')) {
+                    $photo = $request->file('photo');
+                    $photoName = $photo->hashName();
+                    $img = Image::make($photo->getRealPath())->fit(100);
+                    $img->save(public_path(Employee::UPLOAD_PATH).'/'.$photoName);
+
+                    $requestData['uploaded_img'] = $photoName;
+                } else {
+                    $requestData['uploaded_img'] = false;
+                }
 
                 $employeeService->create(collect($requestData), $org, $school);
                 $alert = 'alert_success';
@@ -77,14 +100,28 @@ class EmployeeController extends Controller
 
     public function update(Request $request, EmployeeService $employeeService, OrgService $orgService, Organization $org, Employee $data)
     {
+        if($org->getId() != $data->getOrg()->getId()){
+            return abort(404);
+        }
         if ($request->method() == 'POST') {
-            $request->merge(['org' => $org]);
-            $request->validate([
+            $validation = [
+                'code' => 'required',
                 'name' => 'required',
+                'identity_number' => 'required',
                 'school' => 'required',
-                'org' => 'required',
                 'gender' => 'in:' . Employee::GENDER_MALE . ',' . Employee::GENDER_FEMALE,
-                'dateOfBirth' => 'required|date_format:"d-m-Y'
+                'dateOfBirth' => 'required|date_format:"d-m-Y',
+                'photo' => 'mimes:jpeg,jpg,png,bmp|max:540'
+            ];
+
+            $request->validate($validation, [], [
+                'code' => ucfirst(trans('common.code')),
+                'name' => ucfirst(trans('common.name')),
+                'identity_number' => ucfirst(trans('common.identity_number')),
+                'school' => ucfirst(trans('common.school')),
+                'gender' => ucfirst(trans('common.gender')),
+                'dateOfBirth' => ucfirst(trans('common.date_of_birth')),
+                'photo' => ucfirst(trans('common.photo')),
             ]);
 
             $messageBag = new MessageBag;
@@ -94,12 +131,23 @@ class EmployeeController extends Controller
                 $school = $orgService->findById($request->get('school'));
             }
             if (!$school) {
-                $messageBag->add('school', trans('common.invalid_institute'));
+                $messageBag->add('school', trans('common.invalid_school'));
                 return redirect()->route('administrator.employee.update', ['org' => $org->getId(), 'employee' => $data->getId()])->withErrors($messageBag);
             }
 
             try {
                 $requestData = $request->all();
+
+                if ($request->hasFile('photo')) {
+                    $photo = $request->file('photo');
+                    $photoName = $photo->hashName();
+                    $img = Image::make($photo->getRealPath())->fit(100);
+                    $img->save(public_path(Employee::UPLOAD_PATH).'/'.$photoName);
+
+                    $requestData['uploaded_img'] = $photoName;
+                } else {
+                    $requestData['uploaded_img'] = false;
+                }
 
                 $employeeService->update($data, collect($requestData), false, $school, true);
                 $alert = 'alert_success';
@@ -120,6 +168,9 @@ class EmployeeController extends Controller
 
     public function delete(EmployeeService $employeeService, Organization $org, Employee $data)
     {
+        if($org->getId() != $data->getOrg()->getId()){
+            return abort(404);
+        }
         try {
             $employeeService->delete($data);
             $alert = 'alert_success';
@@ -135,6 +186,9 @@ class EmployeeController extends Controller
 
     public function ajaxDetailEmployee(Request $request, Organization $org, Employee $data)
     {
+        if($org->getId() != $data->getOrg()->getId()){
+            return abort(404);
+        }
         if ($request->ajax()) {
             $data = [
                 'code' => $data->getCode() ? $data->getCode() : '-',
@@ -142,11 +196,12 @@ class EmployeeController extends Controller
                 'school' => ($data->getSchool() instanceof Organization) ? $data->getSchool()->getName() : false,
                 'org' => ($data->getOrg() instanceof Organization) ? $data->getOrg()->getName() : false,
                 'identity_number' => $data->getIdentityNumber() ? $data->getIdentityNumber() : '-',
-                'gender' => $data->getGender() ? $data->getGender() : '-',
+                'gender' => $data->getGender() ? ucfirst($data->getGender()) : '-',
                 'place_of_birth' => $data->getPlaceOfBirth() ? $data->getPlaceOfBirth() : '-',
                 'date_of_birth' => $data->getDateOfBirth() instanceof \DateTime ? $data->getDateOfBirth()->format('d F Y') : '-',
                 'language' => $data->getLanguage() ? $data->getLanguage() : '-',
-                'nationality' => $data->getNationality() ? $data->getNationality() : '-'
+                'nationality' => $data->getNationality() ? $data->getNationality() : '-',
+                'photo' => $data->getPhoto() ? url(url(Employee::UPLOAD_PATH.'/'.$data->getPhoto())) : url('img/avatar.png'),
             ];
 
             return response()->json($data);
