+++ app/Http/Controllers/Administrator/ProgramController.php
@@ -9,6 +9,8 @@ use App\Services\Domain\OrgService;
 use App\Services\Domain\ProgramService;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\MessageBag;
+use Image;
 
 class ProgramController extends Controller
 {
@@ -25,22 +27,28 @@ class ProgramController extends Controller
         $urlDelete = function($id) use ($org) {
             return url(route('administrator.program.delete', [$org->getId(), $id]));
         };
+        $urlDetail = '/org/'.$org->getId().'/program';
 
-        return view('program.index', compact('data', 'page', 'urlCreate', 'urlUpdate', 'urlDelete'));
+        return view('program.index', compact('data', 'page', 'urlCreate', 'urlUpdate', 'urlDelete', 'urlDetail'));
     }
 
     public function create(Request $request, ProgramService $programService, OrgService $orgService, Organization $org)
     {
         if ($request->method() == 'POST') {
-            $request->merge(['org' => $org]);
-            $request->validate([
-                'org' => 'required',
+            $validation = [
                 'name' => 'required',
                 'degree' => 'required|in:'.StudyProgram::DEGREE_D1.','.StudyProgram::DEGREE_D2.','.StudyProgram::DEGREE_D3.','.StudyProgram::DEGREE_S1.','.StudyProgram::DEGREE_S2,
+            ];
+
+            $request->validate($validation, [], [
+                'name' => ucfirst(trans('common.name')),
+                'degree' => ucfirst(trans('common.degree')),
             ]);
 
             try {
-                $programService->create(collect($request->all()));
+                $requestData = $request->all();
+
+                $programService->create(collect($requestData), $org);
                 $alert = 'alert_success';
                 $message = trans('common.create_success', ['object' => ucfirst(trans('common.study_program'))]);
             } catch (Exception $e) {
@@ -57,16 +65,24 @@ class ProgramController extends Controller
 
     public function update(Request $request, ProgramService $programService, OrgService $orgService, Organization $org, StudyProgram $data)
     {
+        if($org->getId() != $data->getOrg()->getId()){
+            return abort(404);
+        }
         if ($request->method() == 'POST') {
-            $request->merge(['org' => $org]);
-            $request->validate([
-                'org' => 'required',
+            $validation = [
                 'name' => 'required',
                 'degree' => 'required|in:'.StudyProgram::DEGREE_D1.','.StudyProgram::DEGREE_D2.','.StudyProgram::DEGREE_D3.','.StudyProgram::DEGREE_S1.','.StudyProgram::DEGREE_S2,
+            ];
+
+            $request->validate($validation, [], [
+                'name' => ucfirst(trans('common.name')),
+                'degree' => ucfirst(trans('common.degree')),
             ]);
 
             try {
-                $programService->update($data, collect($request->input()));
+                $requestData = $request->all();
+
+                $programService->update($data, collect($requestData), false, true);
                 $alert = 'alert_success';
                 $message = trans('common.update_success', ['object' => ucfirst(trans('common.study_program'))]);
             } catch (Exception $e) {
@@ -82,6 +98,9 @@ class ProgramController extends Controller
 
     public function delete(ProgramService $programService, Organization $org, StudyProgram $data)
     {
+        if($org->getId() != $data->getOrg()->getId()){
+            return abort(404);
+        }
         try {
             $programService->delete($data);
             $alert = 'alert_success';
@@ -93,5 +112,24 @@ class ProgramController extends Controller
         }
 
         return redirect()->route('administrator.program.index', ['org' => $org->getId()])->with($alert, $message);
+    }
+
+    public function ajaxDetailProgram(Request $request, Organization $org, StudyProgram $data)
+    {
+        if($org->getId() != $data->getOrg()->getId()){
+            return abort(404);
+        }
+        if ($request->ajax()) {
+            $data = [
+                'code' => $data->getCode() ? $data->getCode() : '-',
+                'name' => $data->getName(),
+                'org' => ($data->getOrg() instanceof Organization) ? $data->getOrg()->getName() : false,
+                'degree' => $data->getDegree() ? ucfirst($data->getDegree()) : '-',
+            ];
+
+            return response()->json($data);
+        }
+
+        return abort(404);
     }
 }
